feat: non-deterministically map services without endpoints
If a service has no endpoints, then the `Port` value in a service will
be used as both the local and remote value. Eventually we'll want to
lookup this port in the service's remap annotation to allow this to be
deterministically set.

We also probably want to eventually document this.